\name{EWDogolny-package}
\alias{EWDogolny-package}
\alias{EWDogolny}
\docType{package}
\title{
Pakiet zawiera ulatwiajace zycie funkcje powstale podczas prac nad modelowaniem wskaznikow EWD.
}
\description{
Funkcje związane są głównie (choć nie tylko) z diagnostyką modeli regresji (w tym mieszanych efektów) estymowanych na dużych danych i eksportem informacji o wyestymowanych modelach poza R.
}
\details{
\tabular{ll}{
Package: \tab EWDogolny\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-13\cr
License: \tab LGPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tomasz Żółtak <tomek@zozlak.org>

Maintainer: Tomasz Żółtak <tomek@zozlak.org>
}
\keyword{ package }

