% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regresja_pierwszego_rodzaju.r
\name{przew_npar_rpr}
\alias{przew_npar_rpr}
\title{Regresja nieparametryczna.}
\usage{
przew_npar_rpr(zmZal, zmNZal, dane, przew_nparPar = NULL)
}
\arguments{
\item{zmZal}{ciąg znaków - nazwa zmiennej zależnej}

\item{zmNZal}{wektor tekstowy z nazwami zmiennych niezależnych}

\item{dane}{ramka danych zawierająca zmienne występujące w formule}

\item{przew_nparPar}{lista z dodatkowymi parametrami dla funkcji przew_npar.}
}
\value{
ramka danych zawierająca zmienne 'x' (wartości zm. niezależnej) i 'y' (przewidywanie)
}
\description{
Funkcja powiela działanie funkcji \code{\link{przew_npar}} i jest uruchamiana z parametrami funkcji \code{\link{regr_pierw_rodz}}.
}

