% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ranef_ddf}
\alias{ranef_ddf}
\title{Predykcje z modeli mieszanych efektow}
\usage{
ranef_ddf(x)
}
\arguments{
\item{x}{model klasy \code{lmerMod}}
}
\value{
lista data frame'ów
}
\description{
Funkcja zwraca predykcje realizacji ef. los. i związane z nimi warunkowe (a'posteriori)
odchylenia standardowe dla poszczególnych grup elegancko połączone w jeden data frame
dla każdego pogrupowania, z którym związane są efekty losowe.
}

