% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regresja_pierwszego_rodzaju.r
\name{regr_pierw_rodz}
\alias{regr_pierw_rodz}
\title{Regresja pierwszego rodzaju.}
\usage{
regr_pierw_rodz(zmZal, zmNZal, dane, FUN = mean)
}
\arguments{
\item{zmZal}{ciąg znaków - nazwa zmiennej zależnej}

\item{zmNZal}{wektor tekstowy z nazwami zmiennych niezależnych}

\item{dane}{ramka danych zawierająca zmienne występujące w formule}

\item{FUN}{funkcja służąca do wyliczeniu parametrów warunkowych}
}
\value{
data frame
}
\description{
Funkcja wylicza regresję pierwszego rodzaju (inaczej mówiąc, parametry warunkowe), domyślnie średnich.
}
\details{
Funkcja \code{FUN} powinna zwracać pojedynczą wartość liczbową i przyjmować argument \code{na.rm}.
}

